advanced_video_generator/cloud_manager.py
"""
Cloud integration module for video generator
"""

import os
import logging
from typing import Optional, Dict, Any
from pathlib import Path

logger = logging.getLogger(__name__)


class CloudManager:
    """Manage cloud storage integrations"""
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        """Initialize cloud manager
        
        Args:
            config: Configuration dictionary
        """
        self.config = config or {}
        self.cloud_config = self.config.get('cloud', {})
        self._drive_mounted = False
    
    def mount_google_drive(self) -> bool:
        """Mount Google Drive in Colab
        
        Returns:
            True if successful
        """
        if self._drive_mounted:
            return True
        
        try:
            from google.colab import drive
            drive.mount('/content/drive')
            self._drive_mounted = True
            logger.info("Google Drive mounted successfully")
            return True
            
        except ImportError:
            logger.warning("Not running in Google Colab")
            return False
        except Exception as e:
            logger.error(f"Failed to mount Google Drive: {e}")
            return False
    
    def save_to_drive(self,
                     local_path: str,
                     drive_path: str,
                     create_folder: bool = True) -> bool:
        """Save a file to Google Drive
        
        Args:
            local_path: Path to local file
            drive_path: Destination path in Google Drive
            create_folder: Whether to create parent folders
            
        Returns:
            True if successful
        """
        try:
            # Ensure drive is mounted
            if not self._drive_mounted:
                if not self.mount_google_drive():
                    return False
            
            # Create parent directory
            if create_folder:
                parent_dir = os.path.dirname(drive_path)
                if parent_dir:
                    Path(parent_dir).mkdir(parents=True, exist_ok=True)
            
            # Copy file
            import shutil
            shutil.copy2(local_path, drive_path)
            
            logger.info(f"Saved to Google Drive: {drive_path}")
            return True
            
        except Exception as e:
            logger.error(f"Failed to save to Google Drive: {e}")
            return False
    
    def upload_file(self,
                   file_path: str,
                   provider: str = "google_drive") -> Optional[str]:
        """Upload a file to cloud storage
        
        Args:
            file_path: Path to file to upload
            provider: Cloud provider (google_drive, dropbox, aws_s3)
            
        Returns:
            URL to uploaded file or None
        """
        if provider == "google_drive":
            return self._upload_to_drive(file_path)
        elif provider == "dropbox":
            return self._upload_to_dropbox(file_path)
        elif provider == "aws_s3":
            return self._upload_to_s3(file_path)
        else:
            logger.warning(f"Unknown provider: {provider}")
            return None
    
    def _upload_to_drive(self, file_path: str) -> Optional[str]:
        """Upload to Google Drive"""
        try:
            if not self._drive_mounted:
                if not self.mount_google_drive():
                    return None
            
            filename = os.path.basename(file_path)
            drive_path = f"/content/drive/MyDrive/VideoGenerator/{filename}"
            
            if self.save_to_drive(file_path, drive_path):
                return drive_path
            return None
            
        except Exception as e:
            logger.error(f"Failed to upload to Drive: {e}")
            return None
    
    def _upload_to_dropbox(self, file_path: str) -> Optional[str]:
        """Upload to Dropbox"""
        try:
            # This would require dropbox package
            # import dropbox
            # Implementation would go here
            logger.warning("Dropbox upload not implemented")
            return None
            
        except Exception as e:
            logger.error(f"Failed to upload to Dropbox: {e}")
            return None
    
    def _upload_to_s3(self, file_path: str) -> Optional[str]:
        """Upload to AWS S3"""
        try:
            # This would require boto3 package
            # import boto3
            # Implementation would go here
            logger.warning("S3 upload not implemented")
            return None
            
        except Exception as e:
            logger.error(f"Failed to upload to S3: {e}")
            return None
    
    def list_drive_files(self, folder: str = "VideoGenerator") -> list:
        """List files in Google Drive folder
        
        Args:
            folder: Folder name
            
        Returns:
            List of file paths
        """
        try:
            if not self._drive_mounted:
                if not self.mount_google_drive():
                    return []
            
            folder_path = f"/content/drive/MyDrive/{folder}"
            if not os.path.exists(folder_path):
                return []
            
            files = []
            for file in Path(folder_path).iterdir():
                if file.is_file():
                    files.append(str(file))
            
            return files
            
        except Exception as e:
            logger.error(f"Failed to list Drive files: {e}")
            return []
    
    def download_from_drive(self,
                           drive_path: str,
                           local_path: str) -> bool:
        """Download a file from Google Drive
        
        Args:
            drive_path: Path in Google Drive
            local_path: Local destination path
            
        Returns:
            True if successful
        """
        try:
            if not self._drive_mounted:
                if not self.mount_google_drive():
                    return False
            
            full_drive_path = f"/content/drive/MyDrive/{drive_path}"
            
            if not os.path.exists(full_drive_path):
                logger.error(f"File not found: {full_drive_path}")
                return False
            
            # Create parent directory
            Path(local_path).parent.mkdir(parents=True, exist_ok=True)
            
            # Copy file
            import shutil
            shutil.copy2(full_drive_path, local_path)
            
            logger.info(f"Downloaded from Drive: {local_path}")
            return True
            
        except Exception as e:
            logger.error(f"Failed to download from Drive: {e}")
            return False


__all__ = ['CloudManager']
