requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# AI/ML
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
scipy>=1.10.0

# Text-to-Speech
gtts>=2.3.0
edge-tts>=6.1.0
pyttsx3>=2.90
TTS>=0.20.0

# Video processing
moviepy>=1.0.3
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Audio processing
pydub>=0.25.1

# Speech recognition (for auto captions)
speechrecognition>=3.10.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Web/UI
flask>=2.3.0
flask-cors>=4.0.0
ipywidgets>=8.0.0
IPython>=8.14.0

# Cloud
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-cloud-storage>=2.10.0

# AWS S3 (optional)
boto3>=1.28.0

# Azure (optional)
# azure-storage-blob>=12.17.0

# Dropbox (optional)
dropbox>=11.36.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.4.0
